+++ efootprint/core/usage/compute_nb_occurrences_in_parallel.py
@@ -21,24 +21,24 @@ def compute_nb_avg_hourly_occurrences(
 
     for hour_shift in range(0, nb_of_full_hours_in_event_duration):
         if nb_avg_hourly_occurrences_in_parallel is None:
-            nb_avg_hourly_occurrences_in_parallel = hourly_occurrences_starts.value.astype(float)
+            nb_avg_hourly_occurrences_in_parallel = hourly_occurrences_starts.value.astype(np.float32, copy=False)
         else:
             nb_avg_hourly_occurrences_in_parallel_padded = np.pad(
-                nb_avg_hourly_occurrences_in_parallel, (0, 1), constant_values=0)
-            shifted_values = np.pad(nb_avg_hourly_occurrences_in_parallel, (1, 0), constant_values=0)
+                nb_avg_hourly_occurrences_in_parallel, (0, 1), constant_values=np.float32(0))
+            shifted_values = np.pad(nb_avg_hourly_occurrences_in_parallel, (1, 0), constant_values=np.float32(0))
             nb_avg_hourly_occurrences_in_parallel = nb_avg_hourly_occurrences_in_parallel_padded + shifted_values
 
     nonfull_duration_rest = event_duration_in_nb_of_hours - nb_of_full_hours_in_event_duration
     if nonfull_duration_rest > 0:
         if nb_avg_hourly_occurrences_in_parallel is None:
             nb_avg_hourly_occurrences_in_parallel = (
-                    hourly_occurrences_starts.value.astype(float) * nonfull_duration_rest)
+                    hourly_occurrences_starts.value.astype(np.float32, copy=False) * nonfull_duration_rest)
         else:
             initial_values_padded = np.pad(
-                nb_avg_hourly_occurrences_in_parallel, (0, 1), constant_values=0)
+                nb_avg_hourly_occurrences_in_parallel, (0, 1), constant_values=np.float32(0))
             shifted_values = np.pad(
-                hourly_occurrences_starts.value.astype(float), (nb_of_full_hours_in_event_duration, 0),
-                constant_values=0)
+                hourly_occurrences_starts.value.astype(np.float32, copy=False), (nb_of_full_hours_in_event_duration, 0),
+                constant_values=np.float32(0))
             nb_avg_hourly_occurrences_in_parallel = initial_values_padded + shifted_values * nonfull_duration_rest
 
     return ExplainableHourlyQuantities(
